+++ docs/docusaurus.config.js
@@ -11,7 +11,7 @@ import "dotenv/config";
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
-  title: "Mastra Documentation",
+  title: "Mastra Docs v1 Beta",
   tagline: "TypeScript agent framework",
   favicon: "/favicon.ico",
 
@@ -27,6 +27,11 @@ const config = {
   markdown: {
     hooks: {
       onBrokenMarkdownLinks: "warn",
+    },
+    parseFrontMatter: async (params) => {
+      const result = await params.defaultParseFrontMatter(params);
+      result.frontMatter.description = `Mastra v1 Beta: ${result.frontMatter.description}`;
+      return result;
     },
   },
   // Enable v4 features in prod