+++ packages/core/src/workflows/types.ts
@@ -9,6 +9,8 @@ import type { ConditionFunction, ExecuteFunction, LoopConditionFunction, Step } 
 export type { ChunkType, WorkflowStreamEvent } from '../stream/types';
 export type { MastraWorkflowStream } from '../stream/MastraWorkflowStream';
 
+export type WorkflowEngineType = string;
+
 export type Emitter = {
   emit: (event: string, data: any) => Promise<void>;
   on: (event: string, callback: (data: any) => void) => void;
@@ -24,6 +26,7 @@ export type StepSuccess<P, R, S, T> = {
   payload: P;
   resumePayload?: R;
   suspendPayload?: S;
+  suspendOutput?: T;
   startedAt: number;
   endedAt: number;
   suspendedAt?: number;
@@ -31,12 +34,13 @@ export type StepSuccess<P, R, S, T> = {
   metadata?: StepMetadata;
 };
 
-export type StepFailure<P, R, S> = {
+export type StepFailure<P, R, S, T> = {
   status: 'failed';
   error: string | Error;
   payload: P;
   resumePayload?: R;
   suspendPayload?: S;
+  suspendOutput?: T;
   startedAt: number;
   endedAt: number;
   suspendedAt?: number;
@@ -44,41 +48,44 @@ export type StepFailure<P, R, S> = {
   metadata?: StepMetadata;
 };
 
-export type StepSuspended<P, S> = {
+export type StepSuspended<P, S, T> = {
   status: 'suspended';
   payload: P;
   suspendPayload?: S;
+  suspendOutput?: T;
   startedAt: number;
   suspendedAt: number;
   metadata?: StepMetadata;
 };
 
-export type StepRunning<P, R, S> = {
+export type StepRunning<P, R, S, T> = {
   status: 'running';
   payload: P;
   resumePayload?: R;
   suspendPayload?: S;
+  suspendOutput?: T;
   startedAt: number;
   suspendedAt?: number;
   resumedAt?: number;
   metadata?: StepMetadata;
 };
 
-export type StepWaiting<P, R, S> = {
+export type StepWaiting<P, R, S, T> = {
   status: 'waiting';
   payload: P;
   suspendPayload?: S;
   resumePayload?: R;
+  suspendOutput?: T;
   startedAt: number;
   metadata?: StepMetadata;
 };
 
 export type StepResult<P, R, S, T> =
   | StepSuccess<P, R, S, T>
-  | StepFailure<P, R, S>
-  | StepSuspended<P, S>
-  | StepRunning<P, R, S>
-  | StepWaiting<P, R, S>;
+  | StepFailure<P, R, S, T>
+  | StepSuspended<P, S, T>
+  | StepRunning<P, R, S, T>
+  | StepWaiting<P, R, S, T>;
 
 export type WorkflowStepStatus = StepResult<any, any, any, any>['status'];
 
@@ -165,6 +172,7 @@ export type ZodPathType<T extends z.ZodTypeAny, P extends string> =
 
 export interface WorkflowState {
   status: WorkflowRunStatus;
+  activeStepsPath: Record<string, number[]>;
   steps: Record<
     string,
     {
@@ -194,7 +202,8 @@ export interface WorkflowRunState {
   value: Record<string, string>;
   context: { input?: Record<string, any> } & Record<string, StepResult<any, any, any, any>>;
   serializedStepGraph: SerializedStepFlowEntry[];
-  activePaths: Array<unknown>;
+  activePaths: Array<number>;
+  activeStepsPath: Record<string, number[]>;
   suspendedPaths: Record<string, number[]>;
   resumeLabels: Record<
     string,