+++ stores/cloudflare/src/storage/index.ts
@@ -261,6 +261,7 @@ export class CloudflareStore extends MastraStorage {
     resourceId,
     fromDate,
     toDate,
+    status,
   }: StorageListWorkflowRunsInput = {}): Promise<WorkflowRuns> {
     return this.stores.workflows.listWorkflowRuns({
       workflowName,
@@ -269,6 +270,7 @@ export class CloudflareStore extends MastraStorage {
       resourceId,
       fromDate,
       toDate,
+      status,
     });
   }